split frequency and vraince thresholding tests
